refactro: Moved warning out to its own subpackage
